# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vetrotech",
        "contact": {
            "email": "P1@zetorzeszow.pl"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/zamowienie": {
            "get": {
                "tags": [
                    "Zamówienie"
                ],
                "summary": "Wyświetl listę zamówień",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zamówienie"
                ],
                "summary": "Dodaj zamówienie",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "orderNumber": {
                                        "type": "string",
                                        "example": "ABC12345"
                                    },
                                    "admissionDate": {
                                        "type": "string",
                                        "example": "2021-01-05"
                                    },
                                    "receiptDate": {
                                        "type": "string",
                                        "example": "2021-01-06"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "contractor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "dimensions_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "width": {
                                        "type": "integer",
                                        "example": "35"
                                    },
                                    "height": {
                                        "type": "integer",
                                        "example": "15"
                                    },
                                    "thickness": {
                                        "type": "integer",
                                        "example": "8"
                                    },
                                    "typeOfGlass_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "nameOfGlass_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "treatment": {
                                        "type": "string",
                                        "example": "Obróbka termiczna"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "numberDepartment_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "comments": {
                                        "type": "string",
                                        "example": "Uwag brak! Uwag brak! Uwag brak! Uwag brak!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "An \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zamowienie/{id}": {
            "get": {
                "tags": [
                    "Zamówienie"
                ],
                "summary": "Wyświetl zamówienie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "An \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zamówienie"
                ],
                "summary": "Edytuj zamówienie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "orderNumber": {
                                        "type": "string",
                                        "example": "ABC12345_EDITED"
                                    },
                                    "admissionDate": {
                                        "type": "string",
                                        "example": "2021-01-05"
                                    },
                                    "receiptDate": {
                                        "type": "string",
                                        "example": "2021-01-06"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "contractor_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "dimensions_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "width": {
                                        "type": "integer",
                                        "example": "35"
                                    },
                                    "height": {
                                        "type": "integer",
                                        "example": "15"
                                    },
                                    "thickness": {
                                        "type": "integer",
                                        "example": "8"
                                    },
                                    "typeOfGlass_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "nameOfGlass_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "treatment": {
                                        "type": "string",
                                        "example": "Obróbka termiczna"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "numberDepartment_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "comments": {
                                        "type": "string",
                                        "example": "Uwag brak! Uwag brak! Uwag brak! Uwag brak!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zamówienie"
                ],
                "summary": "Usuń zamówienie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zamowienie/allProduct": {
            "get": {
                "tags": [
                    "Usunięte zamówienia"
                ],
                "summary": "Lista wszystkich utworzonych zamówień",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zamowienie/deleted": {
            "get": {
                "tags": [
                    "Usunięte zamówienia"
                ],
                "summary": "Lista usuniętych zamówień",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zamowienie/renew/{id}": {
            "get": {
                "tags": [
                    "Usunięte zamówienia"
                ],
                "summary": "Przywróć zamówienie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}